_backup_pre_strict_restore/20260410_074400/requirements.txt
tensorboard
pytorch_lightning
torchmetrics
omegaconf
shortuuid
chumpy
transformers==4.41.2
sentencepiece>=0.1.99
diffusers
einops
wandb
rich
matplotlib

# for visualization
smplx==0.1.28
trimesh==3.9.24
joblib==1.2.0
h5py
scikit-image
spacy
ftfy
more-itertools
natsort
pyrender
moviepy
librosa
triangle==20220202

bpy==3.4.0
numpy==2.1.3
tqdm==4.67.0
pandas
